Fix binary release
